fix some cyclic dependencies
