--- src/vs/base/browser/domImpl/domObservable.ts
@@ -1,17 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { DisposableStore, IDisposable } from '../../common/lifecycle.js';
-import { autorun, IObservable } from '../../common/observable.js';
-import { createStyleSheet2 } from '../domStylesheets.js';
-
-export function createStyleSheetFromObservable(css: IObservable<string>): IDisposable {
-	const store = new DisposableStore();
-	const w = store.add(createStyleSheet2());
-	store.add(autorun(reader => {
-		w.setStyle(css.read(reader));
-	}));
-	return store;
-}
