+++ src/vs/base/browser/domStylesheets.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { DisposableStore, toDisposable, IDisposable } from '../common/lifecycle.js';
+import { autorun, IObservable } from '../common/observable.js';
 import { getWindows, sharedMutationObserver } from './dom.js';
 import { mainWindow } from './window.js';
 
@@ -165,4 +166,13 @@ export function removeCSSRulesContainingSelector(ruleName: string, style = getSh
 
 function isCSSStyleRule(rule: CSSRule): rule is CSSStyleRule {
 	return typeof (rule as CSSStyleRule).selectorText === 'string';
+}
+
+export function createStyleSheetFromObservable(css: IObservable<string>): IDisposable {
+	const store = new DisposableStore();
+	const w = store.add(createStyleSheet2());
+	store.add(autorun(reader => {
+		w.setStyle(css.read(reader));
+	}));
+	return store;
 }
