+++ src/vs/workbench/contrib/chat/common/promptSyntax/codecs/tokens/promptVariable.ts
@@ -4,9 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { PromptToken } from './promptToken.js';
-import { assert } from '../../../../../../../base/common/assert.js';
 import { IRange, Range } from '../../../../../../../editor/common/core/range.js';
-import { INVALID_NAME_CHARACTERS, STOP_CHARACTERS } from '../parsers/promptVariableParser.js';
 
 /**
  * All prompt variables start with `#` character.
@@ -29,14 +27,6 @@ export class PromptVariable extends PromptToken {
 		 */
 		public readonly name: string,
 	) {
-		// sanity check of characters used in the provided variable name
-		for (const character of name) {
-			assert(
-				(INVALID_NAME_CHARACTERS.includes(character) === false) &&
-				(STOP_CHARACTERS.includes(character) === false),
-				`Variable 'name' cannot contain character '${character}', got '${name}'.`,
-			);
-		}
 
 		super(range);
 	}
@@ -74,14 +64,6 @@ export class PromptVariableWithData extends PromptVariable {
 		public readonly data: string,
 	) {
 		super(fullRange, name);
-
-		// sanity check of characters used in the provided variable data
-		for (const character of data) {
-			assert(
-				(STOP_CHARACTERS.includes(character) === false),
-				`Variable 'data' cannot contain character '${character}', got '${data}'.`,
-			);
-		}
 	}
 
 	/**
